unwind the cli config into a package global
I had attempted to keep the CLI configuration as a local varaible but
that wasn't really working out. This unwinds that and leaves the Config
as a package global.

If we ever end up neding to instanciate two CLIs at the same time in the
same process this could pose a problem... we can burn that bridge then.