improve error reporting
We have been ignoring HTTP errors and letting them silently drop on the
floor unless KOSH_DEBUG_MODE was enabled. That's hardly useful in the
general case.

This commit enables passing errors all the way down to the CLI, and adds
errors for when we get 4xx and higher responses from the server.