fix KOSH_URL
So when we added support again for KOSH_ENV we accidentally broken
support for KOSH_URL being set by the environment. The feature we were
relying upon didn't work the way we thought and our tests didn't expose
that properly. This restores the expected behavior that if KOSH_URL is
set that value is taken, otherwise the value of KOSH_ENV determines the
URL used ... or failing that we default to production.

Additonally we add in CONCH_URL, CONCH_TOKEN, and CONCH_ENV aliases for
compatiblity with older/alternative environments.